+++ packages/client/test/end-to-end/decorated-contract.test.ts
@@ -6,7 +6,7 @@ import { toEthereumAddress } from '@streamr/utils'
 import { Contract, JsonRpcProvider, Wallet } from 'ethers'
 import type { StreamRegistryV4 as StreamRegistryContract } from '../../src/ethereumArtifacts/StreamRegistryV4'
 import StreamRegistryArtifact from '../../src/ethereumArtifacts/StreamRegistryV4Abi.json'
-import { createDecoratedContract } from '../../src/utils/contract'
+import { createDecoratedContract } from '../../src/contracts/contract'
 import { mockLoggerFactory } from '../test-utils/utils'
 
 const TEST_CHAIN_CONFIG = CHAIN_CONFIG.dev2
